hwrng: core: add a flush callback
If the current rng is dropped while a read is waiting for data, the cleanup
function cannot be called because a kref is taken for the read, and the new
current rng cannot be called to have data because the reading_mutex is not
released.

This also avoids a hang in hwrng_attr_current_store()
(/sys/class/misc/hw_random/rng_current) because the function ends with an
add_early_randomness() for the new backend that waits on the reading_mutex
owned by the old backend.

Signed-off-by: Laurent Vivier <lvivier@redhat.com>